add backend endpoints

+++ backend/src/main/java/com/codecool/datafilter/dao/model/Data.java
@@ -18,7 +18,7 @@ public class Data {
 
     private String item;
 
-    private int netValue;
+    private double netValue;
 
     private double vat;
